+++ modules/asset/group/tests/src/Kernel/GroupTest.php
@@ -884,4 +884,106 @@ class GroupTest extends KernelTestBase {
     $this->assertCorrectAssets([$asset, $group], $this->assetLocation->getAssetsByLocation([$locations[0]], $timestamps[4] + 1), TRUE);
   }
 
+  /**
+   * Regression test for issue #886.
+   *
+   * Test that creating an asset movement log in the past does not override the
+   * latest group movement log.
+   *
+   * This is a regression test for: Creating old movement logs disrupts current
+   * location logic #886
+   *
+   * @see https://github.com/farmOS/farmOS/issues/886
+   */
+  public function testPastAssetMovement() {
+
+    // Create an animal asset.
+    /** @var \Drupal\asset\Entity\AssetInterface $asset */
+    $asset = Asset::create([
+      'type' => 'animal',
+      'name' => $this->randomMachineName(),
+      'status' => 'active',
+    ]);
+    $asset->save();
+
+    // Create two pasture location assets.
+    /** @var \Drupal\asset\Entity\AssetInterface[] $locations */
+    $locations = [];
+    $location = Asset::create([
+      'type' => 'pasture',
+      'name' => $this->randomMachineName(),
+      'status' => 'active',
+      'intrinsic_geometry' => $this->reduceWkt($this->wktGenerator->wktGeneratePolygon(NULL, rand(3, 7))),
+      'is_fixed' => TRUE,
+      'is_location' => TRUE,
+    ]);
+    $location->save();
+    $locations[] = $location;
+    $location = Asset::create([
+      'type' => 'pasture',
+      'name' => $this->randomMachineName(),
+      'status' => 'active',
+      'intrinsic_geometry' => $this->reduceWkt($this->wktGenerator->wktGeneratePolygon(NULL, rand(3, 7))),
+      'is_fixed' => TRUE,
+      'is_location' => TRUE,
+    ]);
+    $location->save();
+    $locations[] = $location;
+
+    // Create a group asset.
+    /** @var \Drupal\asset\Entity\AssetInterface $group */
+    $group = Asset::create([
+      'type' => 'group',
+      'name' => $this->randomMachineName(),
+      'status' => 'active',
+    ]);
+    $group->save();
+
+    // Assign asset to group.
+    $group_assignment_log = Log::create([
+      'type' => 'test',
+      'status' => 'done',
+      'is_group_assignment' => TRUE,
+      'group' => ['target_id' => $group->id()],
+      'asset' => ['target_id' => $asset->id()],
+    ]);
+    $group_assignment_log->save();
+
+    // Move group to location 1.
+    $group_movement_log = Log::create([
+      'type' => 'test',
+      'status' => 'done',
+      'is_movement' => TRUE,
+      'location' => ['target_id' => $locations[0]->id()],
+      'asset' => ['target_id' => $group->id()],
+    ]);
+    $group_movement_log->save();
+
+    // Confirm that the asset is located where the group is.
+    $this->assertEquals(TRUE, $this->assetLocation->hasLocation($asset));
+    $this->assertEquals(TRUE, $this->assetLocation->hasGeometry($asset));
+    $this->assertEquals($this->logLocation->getLocation($group_movement_log), $this->assetLocation->getLocation($asset));
+    $this->assertEquals($this->logLocation->getGeometry($group_movement_log), $this->assetLocation->getGeometry($asset));
+    $this->assertEquals($group_movement_log->id(), $this->assetLocation->getMovementLog($asset)->id());
+    $this->assertCorrectAssets([$asset, $group], $this->assetLocation->getAssetsByLocation([$locations[0]]), TRUE);
+
+    // Move asset to location 2 with a timestamp of one year ago.
+    $asset_movement_log = Log::create([
+      'timestamp' => strtotime('1 year ago'),
+      'type' => 'test',
+      'status' => 'done',
+      'is_movement' => TRUE,
+      'location' => ['target_id' => $locations[1]->id()],
+      'asset' => ['target_id' => $asset->id()],
+    ]);
+    $asset_movement_log->save();
+
+    // Confirm that the asset is still located where the group is.
+    $this->assertEquals($this->assetLocation->getLocation($group), $this->assetLocation->getLocation($asset));
+    $this->assertEquals($this->assetLocation->getGeometry($group), $this->assetLocation->getGeometry($asset));
+    $this->assertEquals($group_movement_log->id(), $this->assetLocation->getMovementLog($asset)->id());
+    $this->assertCorrectAssets([$asset, $group], $this->assetLocation->getAssetsByLocation([$locations[0]]), TRUE);
+    $this->assertCorrectAssets([], $this->assetLocation->getAssetsByLocation([$locations[1]]), TRUE);
+  }
+
 }